requirements.txt
# Backend API
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.0

# Database
sqlalchemy>=2.0
python-dotenv>=1.0

# Telegram Bot
python-telegram-bot>=20.0
httpx>=0.25

# Password hashing
passlib>=1.7

# PostgreSQL driver
psycopg2-binary>=2.9

# Testing
pytest>=7.0
pytest-asyncio>=0.21
